% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Census2015ByChiefdom}
\alias{Census2015ByChiefdom}
\title{2015 Census Tabulations by Chiefdom}
\format{A data frame with 165 rows and 12 variables:
\describe{
 \item{District}{Name of District}
 \item{CensusChiefdomName}{Name of Chiefdom in Census Tabulation}
 \item{OCHAChiefdomName}{Name of Chiefdom as used by United Nations Office for the Coordination of Humanitarian Affairs}
 \item{TotalPopulation}{Total Population}
 \item{MalePopulation}{Male Population}
 \item{FemalePopulation}{Female Population}
 \item{RuralPopulation}{Rural Population}
 \item{RuralMalePopulation}{Rural Male Population}
 \item{RuralFemalePopulation}{Rural Female Population}
 \item{UrbanPopulation}{Urban Population}
 \item{UrbanMalePopulation}{Urban Male Population}
 \item{UrbanFemalePopulation}{Urban Female Population}
}}
\source{
\url{https://www.statistics.sl/wp-content/uploads/2017/01/final-results_-2015_population_and_housing_census.pdf}
}
\usage{
Census2015ByChiefdom
}
\description{
This data frame contains the Sierra Leone 2015 Census results for selected variables, by Chiefdom
}
\keyword{datasets}
