% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ChiefdomPolygons}
\alias{ChiefdomPolygons}
\title{Spatial polygons for Chiefdom boundaries}
\format{The embedded data frame contains 167 rows (one per Chiefdom) and the following columns:
\describe{
 \item{OBJECTID_1}{}
 \item{OBJECTID}{}
 \item{admin3Name}{Chiefdom Name}
 \item{admin3Pcod}{Chiefdom Code/Identifier}
 \item{admin3RefN}{Chiefdom Reference Name}
 \item{admin2Name}{District Name}
 \item{admin2Pcod}{District Code/Identifier}
 \item{admin2RefN}{District Reference Name}
 \item{admin2AltN}{}
 \item{admin2Al_1}{}
 \item{admin1Name}{Province Name}
 \item{admin1Pcod}{Province Code}
 \item{admin0Name}{Nation name (Sierra Leone)}
 \item{admin0Pcod}{Nation FIPS code}
 \item{date}{Update date}
 \item{validOn}{Validity date}
 \item{ValidTo}{}
 \item{Shape_Leng}{}
 \item{Rowcacode0}{ROWCA Nation code = SLE}
 \item{Rowcacode1}{ROWCA Province code}
 \item{Rowcacode2}{ROWCA District code}
 \item{Shape_Le_1}{}
 \item{Shape_Area}{}
 \item{OCHAChiefdomName}{Recoded Chiefdom name used for merging with other data frames in the slreferencedata package}
 \item{District}{Recoded District name used for merging with other data frames in the slreferencedata package}
 \item{id}{Sequential ID to facilitate merging with fortified data frame}
}}
\source{
\url{https://data.humdata.org/dataset/sierra-leone-all-ad-min-level-boundaries}
}
\usage{
library(ggplot2)
library(dplyr)

chiefdomPolyDf <- fortify(slreferencedata::ChiefdomPolygons)
chiefdomPolyDf <- left_join(chiefdomPolyDf, slreferencedata::ChiefdomPolygons@data) %>%
  left_join(slreferencedata::Census2015ByChiefdom %>% select(District, OCHAChiefdomName, TotalPopulation))

ggplot(data=chiefdomPolyDf, aes(x=long, y=lat, group=group)) +
  geom_path(color="grey") +
  scale_fill_gradient(low = "#ffffcc", high = "#ff4444", space = "Lab", na.value = "grey50", guide = "colourbar", name="Population") +
  geom_polygon(aes(fill=TotalPopulation)) +
  coord_map(projection="mercator")
}
\description{
A SpatialPolygonsDataFrame containing the Chiefdom boundaries in Sierra Leone
Note that ROWCA=OCHA Regional Office for West and Central Africa.
Also note that the embedded data frame has been slightly modified to add two new fields.  The OCHAChiefdomName associates
Tasso Island with the Western Area Urban district and the East III chiefdom, to enable merging with Census and other sources.
The shapefile originally had Tasso Island associated with the Port Loko district, and its own separate chiefdom.
}
\keyword{datasets}
