% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{DistrictPolygons}
\alias{DistrictPolygons}
\title{Spatial polygons for District boundaries}
\format{The embedded data frame contains 14 rows (one per District) and the following columns:
\describe{
 \item{OBJECTID_1}{}
 \item{OBJECTID}{}
 \item{admin2Name}{District Name}
 \item{admin2Pcod}{District Code/Identifier}
 \item{admin2RefN}{District Reference Name}
 \item{admin2AltN}{}
 \item{admin2Al_1}{}
 \item{admin1Name}{Province Name}
 \item{admin1Pcod}{Province Code}
 \item{admin0Name}{Nation name (Sierra Leone)}
 \item{admin0Pcod}{Nation FIPS code}
 \item{date}{Update date}
 \item{validOn}{Validity date}
 \item{ValidTo}{}
 \item{Shape_Leng}{}
 \item{Rowcacode0}{ROWCA Nation code = SLE}
 \item{Rowcacode1}{ROWCA Province code}
 \item{Rowcacode2}{ROWCA District code}
 \item{Shape_Le_1}{}
 \item{Shape_Area}{}
 \item{id}{Sequential ID to facilitate merging with fortified data frame}
}}
\source{
\url{https://data.humdata.org/dataset/sierra-leone-all-ad-min-level-boundaries}
}
\usage{
library(ggplot2)
library(dplyr)

districtPolyDf <- fortify(slreferencedata::DistrictPolygons)
districtPolyDf <- left_join(districtPolyDf, slreferencedata::DistrictPolygons@data) %>%
  left_join(slreferencedata::Census2015ByDistrict %>% select(District, PopTotal), by=c('admin2Name'='District'))

ggplot(data=districtPolyDf, aes(x=long, y=lat, group=group)) +
  geom_path(color="grey") +
  scale_fill_gradient(low = "#ffffcc", high = "#ff4444", space = "Lab", na.value = "grey50", guide = "colourbar", name="Population") +
  geom_polygon(aes(fill=PopTotal)) +
  coord_map(projection="mercator")
}
\description{
A SpatialPolygonsDataFrame containing the District boundaries in Sierra Leone.
Note that ROWCA=OCHA Regional Office for West and Central Africa.
}
\keyword{datasets}
