% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{EbolaCasesWHO}
\alias{EbolaCasesWHO}
\title{Ebola virus cases by week by district}
\format{A data frame with 6,888 rows and 7 variables:
\describe{
 \item{District}{Name of District}
 \item{DataSource}{Source of the data:  Situation Report, or Patient Database}
 \item{CaseDefinition}{Type of case:  Confirmed or Probable}
 \item{WeekLabel}{Label for the week of observation, in WHO format (yyyy-Www, e.g., 2014-W45)}
 \item{Year}{Integer value of the year (2014-2016)}
 \item{Week}{Integer value of the week (1-53)}
 \item{Cases}{Number of Cases}
}}
\source{
\url{http://apps.who.int/gho/data/node.ebola-sitrep.ebola-country-SLE-20160511?lang=en}
}
\usage{
EbolaCasesWHO
}
\description{
This data frame contains the number of Ebola virus cases reported by the World Health Organization
}
\keyword{datasets}
