% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{recodeCensusChiefdomToOCHA}
\alias{recodeCensusChiefdomToOCHA}
\title{This function recodes Chiefdom names as reported by the Sierra Leone census to those used by UN OCHA}
\usage{
recodeCensusChiefdomToOCHA(df, CensusChiefdomNameVariable, DistrictVariable)
}
\arguments{
\item{df}{Data frame with columns for chiefdom name and district name}

\item{CensusChiefdomNameVariable}{The name of the column containing the census chiefdom name}

\item{DistrictVariable}{The name of the column containing the census district name}
}
\value{
The data frame, with a new column (OCHAChiefdomName) added
}
\description{
This function recodes Chiefdom names as reported by the Sierra Leone census to those used by UN OCHA
}
